Fix typo for test float64

+++ tests/test_basics.py
@@ -48,7 +48,7 @@ def test_float64():
     assert str(a) == "Float64[1.20000001, 2.2]"
     assert a.type == ad.DataType.float64()
 
-    a = ad.Float32Array([1.20000001, None])
+    a = ad.Float64Array([1.20000001, None])
     assert str(a) == "Float64[1.20000001, None]"
     assert len(a) == 2
     assert list(a) == [1.2, None]
